@@ -1,3 +1,5 @@
 module github.com/ContinuumLLC/cpuid
 
-go 1.12
+go 1.15
+
+require golang.org/x/sys v0.5.0
